refactor(app): extract NDDScore table controller (#346)
Move load, URL-sync, filter/pagination state, and action handlers out of
NddScoreGeneTable.vue (822 lines) into useNddScoreGeneTable.ts so the
component stays a thin presentation shell. Existing filter-construction,
filter-UI, and column modules are reused unchanged. Behavior-preserving;
NDDScore ML-prediction labelling/wording is untouched.

Adds useNddScoreGeneTable.spec.ts covering URL hydration (range/HPO/search/
sort/page/page-size), monotonic request-serial rejection of stale success
and stale error responses, full filter reset, and graceful degradation when
the HPO term lookup fails.
