+++ app/src/components/nddscore/NddScoreGeneTable.vue
@@ -323,7 +323,6 @@
 </template>
 
 <script setup lang="ts">
-import { computed, onMounted, reactive, ref } from 'vue';
 import { RouterLink } from 'vue-router';
 import {
   BAlert,
@@ -339,7 +338,6 @@ import {
   BFormSelect,
   BRow,
 } from 'bootstrap-vue-next';
-import { fetchGenePredictions, fetchHpoTerms, type NddScoreGenePrediction } from '@/api/nddscore';
 import GeneBadge from '@/components/ui/GeneBadge.vue';
 import TableShell from '@/components/table/TableShell.vue';
 import GenericTable from '@/components/small/GenericTable.vue';
@@ -347,21 +345,6 @@ import NddScoreGeneMobileRows from './NddScoreGeneMobileRows.vue';
 import TableSearchInput from '@/components/small/TableSearchInput.vue';
 import TablePaginationControls from '@/components/small/TablePaginationControls.vue';
 import TableDownloadLinkCopyButtons from '@/components/small/TableDownloadLinkCopyButtons.vue';
-import { useExcelExport } from '@/composables/useExcelExport';
-import { withReturnTo } from '@/utils/returnNavigation';
-import {
-  buildNddScoreGeneApiFilters,
-  buildNddScoreGeneFilterString,
-  parseNddScoreGeneFilterClauses,
-  type NddScoreGeneRangeOperator,
-} from './nddScoreGeneTableFilters';
-import {
-  nddScoreGeneFields,
-  nddScoreGeneColumnHelp,
-  nddScoreRangeOperatorOptions,
-  type NddScoreGeneFieldDefinition,
-  type NddScoreGeneFieldKey,
-} from './nddScoreGeneTableColumns';
 import {
   displayValue,
   formatDecimal,
@@ -370,16 +353,7 @@ import {
   topHpoLabel,
   topHpoTooltip,
 } from './nddScoreGeneTableFormatters';
-import {
-  nddScoreSelectOptionsFor,
-  nddScoreRangeValuePlaceholder,
-  nddScoreRangeFilterLabel,
-  nddScoreFilterDropdownToggleClass,
-  nddScoreFilterDropdownClass,
-  nddScoreFilterControlClass,
-  nddScoreHpoTermOption,
-  type NddScoreSelectOption,
-} from './nddScoreGeneTableFilterUi';
+import { useNddScoreGeneTable } from './useNddScoreGeneTable';
 
 defineOptions({
   name: 'NddScoreGeneTable',
@@ -414,409 +388,52 @@ function confidenceChipClass(value: unknown): string {
   }
 }
 
-type GenePredictionRow = NddScoreGenePrediction & {
-  release_id?: string;
-  hgnc_id?: string | number;
-  gene_symbol?: string;
-  ndd_score?: number | string;
-  rank?: number | string;
-  percentile?: number | string;
-  risk_tier?: string;
-  confidence_tier?: string;
-  known_sysndd_gene?: boolean | number | string | null;
-  model_split?: string | null;
-  top_inheritance_mode?: string | null;
-  n_predicted_hpo?: number | string | null;
-  top_hpo_predictions_json?: unknown;
-};
-
-type SortEvent = {
-  sortBy: string;
-  sortDesc: boolean;
-};
-
-type FieldKey = NddScoreGeneFieldKey;
-type FieldDefinition = NddScoreGeneFieldDefinition;
-
-type RangeOperator = NddScoreGeneRangeOperator;
-type RangeFieldKey = 'ndd_score' | 'rank' | 'percentile';
-type RangeFilterState = {
-  operator: RangeOperator;
-  value: string;
-  valueMax: string;
-};
-type UrlFilterClause = {
-  operator: string;
-  key: string;
-  values: string[];
-};
-
-const rows = ref<GenePredictionRow[]>([]);
-const total = ref(0);
-const page = ref(1);
-const pageSize = ref(10);
-const loading = ref(false);
-const hasLoadedOnce = ref(false);
-const loadError = ref('');
-const search = ref('');
-const sort = ref('rank');
-const hpoTermFilter = ref<string[]>([]);
-const hpoTermOptions = ref<NddScoreSelectOption[]>([]);
-const hpoTermSearch = ref('');
-let requestSerial = 0;
-const { isExporting, exportToExcel } = useExcelExport();
-
-const rangeOperatorOptions = nddScoreRangeOperatorOptions;
-const fields: FieldDefinition[] = nddScoreGeneFields;
-
-const columnFilters = reactive<Record<string, string>>({
-  gene_symbol: '',
-  ndd_score_min: '',
-  ndd_score_max: '',
-  rank_min: '',
-  rank_max: '',
-  percentile_min: '',
-  percentile_max: '',
-  risk_tier: '',
-  confidence_tier: '',
-  known_sysndd_gene: '',
-  model_split: '',
-  top_inheritance_mode: '',
-});
-
-const rangeFilters = reactive<Record<RangeFieldKey, RangeFilterState>>({
-  ndd_score: { operator: 'any', value: '', valueMax: '' },
-  rank: { operator: 'any', value: '', valueMax: '' },
-  percentile: { operator: 'any', value: '', valueMax: '' },
-});
-
-const columnHelp = nddScoreGeneColumnHelp;
-
-const totalLabel = computed(() => `${total.value.toLocaleString()} genes`);
-const tableShellLoading = computed(() => loading.value && !hasLoadedOnce.value);
-const hasActiveFilters = computed(
-  () =>
-    search.value.trim() !== '' ||
-    Object.values(columnFilters).some((value) => value.trim() !== '') ||
-    Object.values(rangeFilters).some((state) => state.operator !== 'any') ||
-    hpoTermFilter.value.length > 0
-);
-const sortBy = computed(() => [
-  {
-    key: sort.value.replace(/^[+-]/, ''),
-    order: sort.value.startsWith('-') ? 'desc' : 'asc',
-  },
-]);
-const filteredHpoTermOptions = computed(() => {
-  const query = hpoTermSearch.value.trim().toLowerCase();
-  if (!query) {
-    return hpoTermOptions.value;
-  }
-  return hpoTermOptions.value.filter(
-    (option) =>
-      option.value.toLowerCase().includes(query) || option.text.toLowerCase().includes(query)
-  );
-});
-const hpoFilterLabel = computed(() => {
-  if (hpoTermFilter.value.length === 0) {
-    return 'Any HPO';
-  }
-  if (hpoTermFilter.value.length === 1) {
-    return hpoTermFilter.value[0];
-  }
-  return `${hpoTermFilter.value.length} HPO terms`;
-});
-
-const hpoFilterToggleClass = computed(() =>
-  nddScoreFilterDropdownToggleClass(hpoTermFilter.value.length === 0)
-);
-
-const hpoFilterDropdownClass = computed(() =>
-  nddScoreFilterDropdownClass(hpoTermFilter.value.length === 0)
-);
-
-function normalizeRows(data: NddScoreGenePrediction[] | undefined): GenePredictionRow[] {
-  return (data ?? []).map((row) => row as GenePredictionRow);
-}
-
-async function loadHpoTermOptions() {
-  try {
-    hpoTermOptions.value = (await fetchHpoTerms())
-      .map(nddScoreHpoTermOption)
-      .filter((option): option is NddScoreSelectOption => option != null);
-  } catch {
-    hpoTermOptions.value = [];
-  }
-}
-
-function selectOptionsFor(field: FieldDefinition): NddScoreSelectOption[] {
-  return nddScoreSelectOptionsFor(field);
-}
-
-function rangeKey(key: FieldKey): RangeFieldKey {
-  return key as RangeFieldKey;
-}
-
-function rangeValuePlaceholder(field: FieldDefinition): string {
-  return nddScoreRangeValuePlaceholder(rangeFilters[rangeKey(field.key)].operator);
-}
-
-function rangeFilterLabel(field: FieldDefinition): string {
-  return nddScoreRangeFilterLabel(field, rangeFilters[rangeKey(field.key)]);
-}
-
-function rangeFilterToggleClass(field: FieldDefinition): string {
-  return nddScoreFilterDropdownToggleClass(rangeFilters[rangeKey(field.key)].operator === 'any');
-}
-
-function rangeFilterDropdownClass(field: FieldDefinition): Record<string, boolean> {
-  return nddScoreFilterDropdownClass(rangeFilters[rangeKey(field.key)].operator === 'any');
-}
-
-function filterControlClass(key: FieldKey): Record<string, boolean> {
-  return nddScoreFilterControlClass(!columnFilters[key]);
-}
-
-function handleRangeOperatorChange(key: FieldKey) {
-  const state = rangeFilters[rangeKey(key)];
-  if (state.operator === 'any') {
-    state.value = '';
-    state.valueMax = '';
-    handleColumnFilterChange();
-  } else if (state.operator !== 'range') {
-    state.valueMax = '';
-  }
-  removeSearch();
-}
-
-function clearRangeFilter(key: FieldKey) {
-  const state = rangeFilters[rangeKey(key)];
-  state.operator = 'any';
-  state.value = '';
-  state.valueMax = '';
-  handleColumnFilterChange();
-}
-
-function toggleHpoTerm(value: string) {
-  if (hpoTermFilter.value.includes(value)) {
-    hpoTermFilter.value = hpoTermFilter.value.filter((term) => term !== value);
-  } else {
-    hpoTermFilter.value = [...hpoTermFilter.value, value];
-  }
-  removeSearch();
-  handleColumnFilterChange();
-}
-
-function clearHpoTerms() {
-  hpoTermFilter.value = [];
-  hpoTermSearch.value = '';
-  handleColumnFilterChange();
-}
-
-async function requestExcel() {
-  await exportToExcel(rows.value, {
-    filename: 'nddscore_gene_predictions',
-    sheetName: 'NDDScore genes',
-  });
-}
-
-async function copyLinkToClipboard() {
-  await navigator.clipboard?.writeText(window.location.href);
-}
-
-function removeFilters() {
-  search.value = '';
-  Object.keys(columnFilters).forEach((key) => {
-    columnFilters[key] = '';
-  });
-  (Object.keys(rangeFilters) as RangeFieldKey[]).forEach((key) => {
-    rangeFilters[key].operator = 'any';
-    rangeFilters[key].value = '';
-    rangeFilters[key].valueMax = '';
-  });
-  hpoTermFilter.value = [];
-  hpoTermSearch.value = '';
-  page.value = 1;
-  void loadRows();
-}
-
-function activeFilterString(): string {
-  return buildNddScoreGeneFilterString({
-    search: search.value,
-    columnFilters,
-    rangeFilters,
-    hpoTerms: hpoTermFilter.value,
-  });
-}
-
-function normalizedSortForUrl(): string {
-  const sortKey = sort.value.replace(/^[+-]/, '');
-  return sort.value.startsWith('-') ? `-${sortKey}` : `+${sortKey}`;
-}
-
-function normalizedSortForApi(): string {
-  const sortKey = sort.value.replace(/^[+-]/, '');
-  return sort.value.startsWith('-') ? `-${sortKey}` : sortKey;
-}
-
-function updateBrowserUrl() {
-  if (typeof window === 'undefined') {
-    return;
-  }
-
-  const params = new URLSearchParams();
-  params.set('sort', normalizedSortForUrl());
-  const filter = activeFilterString();
-  if (filter) {
-    params.set('filter', filter);
-  }
-  if (page.value > 1) {
-    params.set('page', String(page.value));
-  }
-  params.set('page_size', String(pageSize.value));
-
-  const query = params.toString();
-  const nextUrl = query ? `${window.location.pathname}?${query}` : window.location.pathname;
-  window.history.replaceState({ ...window.history.state }, '', nextUrl);
-}
-
-function parseFilterClauses(filterString: string | null): UrlFilterClause[] {
-  return parseNddScoreGeneFilterClauses(filterString);
-}
-
-function applyInitialQuery() {
-  if (typeof window === 'undefined') {
-    return;
-  }
-
-  const params = new URLSearchParams(window.location.search);
-  sort.value = params.get('sort') || sort.value;
-  search.value = '';
-  page.value = Math.max(1, Number(params.get('page')) || 1);
-  pageSize.value = Math.max(1, Number(params.get('page_size')) || 10);
-  Object.keys(columnFilters).forEach((key) => {
-    columnFilters[key] = '';
-  });
-  (Object.keys(rangeFilters) as RangeFieldKey[]).forEach((key) => {
-    rangeFilters[key].operator = 'any';
-    rangeFilters[key].value = '';
-    rangeFilters[key].valueMax = '';
-  });
-  hpoTermFilter.value = [];
-
-  parseFilterClauses(params.get('filter')).forEach((clause) => {
-    const [first = '', second = ''] = clause.values;
-    if (clause.key === 'any') {
-      search.value = first;
-      return;
-    }
-    if (clause.key in columnFilters) {
-      columnFilters[clause.key] = first;
-      return;
-    }
-    if (clause.key === 'top_hpo_predictions_json') {
-      hpoTermFilter.value = clause.values;
-      return;
-    }
-    if (clause.key in rangeFilters) {
-      const key = clause.key as RangeFieldKey;
-      if (clause.operator === 'range') {
-        rangeFilters[key].operator = 'range';
-        rangeFilters[key].value = first;
-        rangeFilters[key].valueMax = second;
-      } else if (['gte', 'lte', 'eq'].includes(clause.operator)) {
-        rangeFilters[key].operator = clause.operator as RangeOperator;
-        rangeFilters[key].value = first;
-      }
-    }
-  });
-}
-
-async function loadRows() {
-  const serial = ++requestSerial;
-  loading.value = true;
-  loadError.value = '';
-
-  try {
-    const apiFilters = buildNddScoreGeneApiFilters({
-      search: search.value,
-      columnFilters,
-      rangeFilters,
-      hpoTerms: hpoTermFilter.value,
-    });
-
-    const result = await fetchGenePredictions({
-      sort: normalizedSortForApi(),
-      ...apiFilters,
-      page: page.value,
-      pageSize: pageSize.value,
-      hgncId: undefined,
-    });
-
-    if (serial !== requestSerial) {
-      return;
-    }
-
-    rows.value = normalizeRows(result.data);
-    total.value = Number(result.total) || 0;
-    page.value = Number(result.page) || page.value;
-    pageSize.value = Number(result.page_size) || pageSize.value;
-    hasLoadedOnce.value = true;
-    updateBrowserUrl();
-  } catch {
-    if (serial === requestSerial) {
-      rows.value = [];
-      total.value = 0;
-      loadError.value = 'NDDScore predictions are not available for the active release.';
-      hasLoadedOnce.value = true;
-    }
-  } finally {
-    if (serial === requestSerial) {
-      loading.value = false;
-    }
-  }
-}
-
-function handleSearchChange() {
-  page.value = 1;
-  void loadRows();
-}
-
-function handleColumnFilterChange() {
-  page.value = 1;
-  void loadRows();
-}
-
-function handleSortUpdate({ sortBy: nextSortBy, sortDesc }: SortEvent) {
-  sort.value = `${sortDesc ? '-' : ''}${nextSortBy}`;
-  page.value = 1;
-  void loadRows();
-}
-
-function handlePageChange(nextPage: number) {
-  page.value = nextPage;
-  void loadRows();
-}
-
-function handlePageSizeChange(nextPageSize: number) {
-  pageSize.value = nextPageSize;
-  page.value = 1;
-  void loadRows();
-}
-
-function removeSearch() {
-  search.value = '';
-}
-
-function detailPath(row: GenePredictionRow): string {
-  return withReturnTo(`/NDDScore/Gene/${encodeURIComponent(displayValue(row.hgnc_id))}`);
-}
-
-onMounted(() => {
-  applyInitialQuery();
-  void loadHpoTermOptions();
-  void loadRows();
-});
+const {
+  rows,
+  total,
+  page,
+  pageSize,
+  loading,
+  loadError,
+  search,
+  hpoTermFilter,
+  hpoTermSearch,
+  isExporting,
+  rangeOperatorOptions,
+  fields,
+  columnFilters,
+  rangeFilters,
+  columnHelp,
+  totalLabel,
+  tableShellLoading,
+  hasActiveFilters,
+  sortBy,
+  filteredHpoTermOptions,
+  hpoFilterLabel,
+  hpoFilterToggleClass,
+  hpoFilterDropdownClass,
+  selectOptionsFor,
+  rangeKey,
+  rangeValuePlaceholder,
+  rangeFilterLabel,
+  rangeFilterToggleClass,
+  rangeFilterDropdownClass,
+  filterControlClass,
+  handleRangeOperatorChange,
+  clearRangeFilter,
+  toggleHpoTerm,
+  clearHpoTerms,
+  requestExcel,
+  copyLinkToClipboard,
+  removeFilters,
+  handleSearchChange,
+  handleColumnFilterChange,
+  handleSortUpdate,
+  handlePageChange,
+  handlePageSizeChange,
+  removeSearch,
+  detailPath,
+} = useNddScoreGeneTable();
 </script>
 
 <style scoped src="./NddScoreGeneTable.styles.css"></style>
